Update CalendlyPopupWidget.vue
className in the nested divs in this template breaks the styling on this badge. The root className is converted to class, but the nested className remains className and fails to convert correctly, so the class is not applied, and the style breaks. 

Changing both instances of className to class successfully resolves the issue and conforms to the official Nuxt docs.

(Thanks Fabian for building this time-saving module.)

Related issue: https://github.com/nuxt/nuxt/discussions/24548 | Nested children className not rendered into class when refreshing in SSR mode #24548

Working syntax documented here: https://vuejs.org/guide/essentials/class-and-style